add TootClient.getFeaturedTagsSuggestions()
accesses Mastodon featured_tags suggestions API

also moved existing getFeaturedTags into new file TootClient+FeaturedTags
