+++ Sources/TootSDK/TootClient/TootClient+Account.swift
@@ -85,18 +85,6 @@ extension TootClient {
         }
     }
 
-    /// Get tags featured by user.
-    ///
-    /// - Parameter userID: ID of user in database.
-    /// - Returns: The featured tags or an error if unable to retrieve.
-    public func getFeaturedTags(forUser userID: String) async throws -> [FeaturedTag] {
-        let req = HTTPRequestBuilder {
-            $0.url = getURL(["api", "v1", "accounts", userID, "featured_tags"])
-            $0.method = .get
-        }
-        return try await fetch([FeaturedTag].self, req)
-    }
-
     // swiftlint:disable todo
     // TODO: - Update account credentials
